projects/project-popcorn/requirements-prod.txt
# Cross-Domain Radar Production Dependencies
# Generated: 2026-01-19

# LLM & Embedding
openai>=2.15.0

# Data Processing
pandas>=2.3.0
numpy>=2.4.0
pyyaml>=6.0.0

# HTTP & API
requests>=2.32.0
httpx>=0.28.0

# Utilities
python-dotenv>=1.2.0
tqdm>=4.67.0

# Optional: Visualization (for reports)
matplotlib>=3.10.0
scikit-learn>=1.3.0
